Refactoring: Change Navbar component to typescript
